Implement prepend patch for postgres
We've had issues running rack-mini-profiler alongside dd-trace (see:
https://github.com/DataDog/dd-trace-rb/issues/2348).

The suggested solution is to add support for prepend style patching for
pg, similar to the one that exists for mysql. This does that.
