+++ lib/prepend_pg_patch.rb
@@ -1,0 +1,5 @@
+# frozen_string_literal: true
+
+module Rack
+  MINI_PROFILER_PREPEND_PG_PATCH = true
+end
